pem: refactor to expose pem iterator privately
This refactors pem.c to expose a new internal API, pem_next. This
API will allow iteration/parsing of PEM(s). Other APIs in pem.c were
updated to use pem_next.

pem_next behaves similar to pem_load_buffer, except it does not do
any base64 decoding. Instead it just returns the start of the PEM as
a string, and the length as an out parameter. This allows simple
PEM iteration loops. Strictly for pem.c usage we decode the base64
string in the loop, but settings.c will need this for parsing PEMs
from settings files and returning the PEM as a full string, including
the start/end tags.

endp was removed from pem_load_buffer as it was no longer being used.